requirements.txt
git+https://github.com/Pycord-Development/pycord
pickledb
python-dotenv